Changed is now a property instead of a method. Objects are now loaded/displayed in-game. BUT they're not 100% correct; not sure why.

+++ src/OpenClassic.Server/Domain/Definition/GameObjectDefinition.cs
@@ -18,7 +18,7 @@
         public short Id { get; set; }
         public short X { get; set; }
         public short Y { get; set; }
-        public int Direction { get; set; }
+        public sbyte Direction { get; set; }
         public int Type { get; set; }
     }
 }
